Adds create_annotation_from_segmentation_map
This is a greatly reduced version from the OTX implementation.
The OTX implementation has currently some functionality with hole
handling.
So far, with soft prediction this is not relevant, but it might be later.
The hard prediction also was created with a size which needs to be
created { cols, rows }. This was a bug that caused the hard prediction
to be partly undefined.