Prepare for YoloV8
Add reverse_input_channels which has a value of true for YoloV8.

Add scale_values which has a value of [255] for YoloV8.

confidenceThreshold->confidence_threshold to align with python.
If we are going to follow OpenVINO's code style this is also a good change.

boxIOUThreshold->iou_threshold, iou_t->iou_threshold:
Everything that ends with _t is treated as type in my IDE.
This also aligns with confidence_threshold.

auto->const ov::AnyMap& I was in doubts if the copy happens.
Now the copy doesn't happen for shure.

Call testing::InitGoogleTest(&argc, argv) first.
gtests understands some cmd args. Let it see them.