Segmentation CPP generate annotations (#57)
* Adds create_annotation_from_segmentation_map

This is a greatly reduced version from the OTX implementation.
The OTX implementation has currently some functionality with hole
handling.
So far, with soft prediction this is not relevant, but it might be later.
The hard prediction also was created with a size which needs to be
created { cols, rows }. This was a bug that caused the hard prediction
to be partly undefined.

* Refactor the post processing

The output tensor can be one channel
If it is one channel, check if its 32 or 64bit and just set the hard_prediction with that. No blurring, I guess ( check ).

If there are mulitple channels:
- Blur if blur_strength != -1 ( or > 0 )
- Per pixel get the channel with the highest probability. and write to hard_prediciton map

Then use that hard prediction map to get the annotations.

* Add sanity test for segmentation

* Apply feedback and fixes tests

Fighting with the formatter, ignore the format. Will fix tomorrow

* Format segmentation model

* Fix error in getting label

Background seems to not be defined in the labels.
That means the index is off by one for the label.

* Make the softThreshold conform to documentation

* Fix tests

* WIP

* wip

* getContours working

* Revert the changes to synchronous api example

* move get_contours python version as method

* Fixed undefined result without return_soft_prediction

* Fixes tests

Python precommit test wasn't handling the tuples
Fixed sanity cpp test using old contour implementation

* Remove newline

* Change interface to use reference instead of shared pointer.

The function just uses data from the imageResult

* Fix confusing naming in hard_prediction creation

* Return empty contour when calling get_contours on model without labels

* Set -inf as soft_threshold default

* Revert the default soft_threshold value

* Move the resize so that blur happens first, and then resize

* Apply feedback for astype

* Reformat py

* Remove clone in create_hard_prediction

* Remove double transpose

* Go back to default soft_threshold of -infinity and fix bug with or on blurring

I'm unsure if this is the correct default.
Aside from that there was a bug in applying the blur in python.
It was and, and it should've been or

* Fix broken sanity test

* CPP tests for contours

* Add more references for segments to the public scope

* Working tests with contours

* Move contour struct to results.h

* Fix tests for contours

* Fix style change.

* Apply feedback...

* Remove automatically added change by format

* Apply black...

* Reverse blur and resize in python too

* Throws runtime error when trying to get contours from soft prediction with one layer